feat: added gm8126 carving handler
